requirements.txt
# Core scientific computing
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# EEG processing
mne>=1.0.0
mne-bids>=0.10.0

# Deep learning
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0

# MLOps and experiment tracking
wandb>=0.13.0
tensorboard>=2.8.0

# Data visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.64.0
scikit-learn>=1.1.0

# Optional: Jupyter for notebooks
jupyter>=1.0.0
ipykernel>=6.0.0

# Optional: Docker support
docker>=6.0.0 